refactor(shared): move validate into validate-core so process.exit stays out of the browser graph
Mirrors the generate-core/generate and diff-core/diff seam: validate.ts kept
exitBasedOffOfValidationOutcome (three process.exit calls) inside the module
graph reachable from browser.ts. Moving validate() and its helpers into
validate-core.ts lets browser.ts import the pure core directly, dropping
exitBasedOffOfValidationOutcome out of the browser bundle entirely.
