+++ shared/AGENTS.md
@@ -47,8 +47,8 @@ npx vitest run ${TEST FILE}
 
 Rules:
 - New modules are browser-safe by default. Node-only code (`fs`, `path`, `net`, `process.exit`, `__dirname`, winston, mkdirp, playwright) lives in a `*.node.ts` / `node-*.ts` module or in a wrapper that the root barrel imports — never imported from `browser.ts` or anything it reaches.
-- Prefer seams over conditionals: pure core + Node wrapper (`generate-core.ts` / `generate.ts`, `diff-core.ts` / `diff.ts`), injected `DocumentLoader`s, registries (`registerNodeLoggerFactory`, `registerOutputFormatter`).
-- `scripts/check-browser-entry.mjs` runs in `npm test`. It bundles `src/browser.ts` with esbuild for the browser and fails on any Node builtin request outside a four-entry allowlist (Spectral's dependency chain requests `fs`/`path`/`buffer` but never touches `fs`/`path` at runtime), then executes a real `validate()` probe with those builtins stubbed to throw. Do not extend the allowlist to make a red build green — fix the seam.
+- Prefer seams over conditionals: pure core + Node wrapper (`generate-core.ts` / `generate.ts`, `diff-core.ts` / `diff.ts`, `validate-core.ts` / `validate.ts`), injected `DocumentLoader`s, registries (`registerNodeLoggerFactory`, `registerOutputFormatter`).
+- `scripts/check-browser-entry.mjs` runs in `npm test`. It bundles `src/browser.ts` with esbuild for the browser and fails on any Node builtin request outside a four-entry allowlist (Spectral's dependency chain requests `fs`/`path`/`buffer` but never touches `fs`/`path` at runtime), then executes a real probe (`validate()`, `generate()`, `diffDocuments()`) with those builtins stubbed to throw. Do not extend the allowlist to make a red build green — fix the seam.
 - Deep imports (`@finos/calm-shared/src/...`, `/dist/...`) are sealed by the `exports` map. Import from the barrel.
 - Browser consumers bundling the entry must map the allowlisted builtins to nothing — webpack: `resolve.fallback: { fs: false, path: false, buffer: false }`; esbuild: the same stub plugin the guard uses.
 - Not in the browser entry (follow-ups): template/docify (filesystem-bound loaders and output strategies), Hub read/write commands (CORS), diagram rasterisation.
